+++ src/utils/styleVariables.js
@@ -1,0 +1,8 @@
+export const colors = {
+  black: '#000000',
+  white: '#FFFFFF',
+  gray900: '#323232',
+  gold: '#998100',
+};
+
+export default colors;